% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_functions.R
\name{learn_lists}
\alias{learn_lists}
\title{Run the learning lists tutorial}
\usage{
learn_lists()
}
\value{

}
\description{
Run the learning lists tutorial
}
\examples{
if(is.interactive()){
  learn_lists()
}
}
