% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_functions.R
\name{learn_purrr}
\alias{learn_purrr}
\title{Run the learning purr tutorial}
\usage{
learn_purrr()
}
\value{

}
\description{
Run the learning purr tutorial
}
\examples{
if(is.interactive()){
  learn_purrr()
}
}
