% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_functions.R
\name{learn_tidyselect}
\alias{learn_tidyselect}
\title{Run the learning tidyselect tutorial}
\usage{
learn_tidyselect()
}
\value{

}
\description{
Run the learning tidyselect tutorial
}
\examples{
if(is.interactive()){
learn_tidyselect()
}

}
